Rich text: extract indent list item on space
